requirements.txt
annotated-types==0.7.0
blinker==1.9.0
blis==1.2.0
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
cymem==2.0.11
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
filelock==3.17.0
Flask==3.1.0
fsspec==2025.2.0
huggingface-hub==0.29.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
langcodes==3.5.0
language_data==1.3.0
llvmlite==0.44.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.6.0
mpmath==1.3.0
murmurhash==1.0.12
networkx==3.4.2
nltk==3.9.1
numba==0.61.0
numpy==2.1.3
openai-whisper==20240930
packaging==24.2
preshed==3.0.9
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.3
shellingham==1.5.4
smart-open==7.1.0
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
summarizer==0.0.7
sympy==1.13.1
thinc==8.3.4
tiktoken==0.9.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
typer==0.15.2
typing_extensions==4.12.2
urllib3==2.3.0
wasabi==1.1.3
weasel==0.4.1
Werkzeug==3.1.3
wrapt==1.17.2
